sanitised queue_pop, queue_write fixed problem with queue_delete (managerLen wasn't being updated and all queues from queue_manager above the deleted one did not get pushed down. queue_del_update(pos,cursor) fixes that)

+++ api_core.py
@@ -3,6 +3,21 @@ import time
 import random, string
 def san_bool(str:str):
     return (("'" not in str) and ('"' not in str) and (";" not in str) and ("*" not in str))
+
+def length_update(name="queue"):
+    res = -1 #error state
+    if san_bool(name):
+        with sq3.connect("queue.db") as database:
+            c = database.cursor()
+            c.execute(f"select * from {name}")
+            res = len(c.fetchall())
+            c.execute("update queue_manager set len = ? where name = ?",(res,str(name)))
+            database.commit()
+            c.close()
+    else:
+        print(f"length_update received unsafe name {name}")
+    return res
+
 with sq3.connect("queue.db") as database:
     PROTECTED_NAMES = ("queue", "info_queue", "info_event", "queue_manager")
     def queue_create_new(name="queue"):
@@ -56,9 +71,11 @@ with sq3.connect("queue.db") as database:
     if len(c.fetchmany(1)) == 0:
         c.execute(f"{MNG_INSERT} ({managerLen}, 0, 'info_events', '', '', '', {round(time.time())})")
         managerLen += 1
-
+    length_update("info_queue")
+    length_update("info_events")
+    length_update("queue")
     c.execute("select * from queue_manager")
-    print(c.fetchall())
+    #print(c.fetchall())
     database.commit()
     c.close()
 
@@ -70,52 +87,50 @@ with sq3.connect("queue.db") as database:
     database.commit()
     c.close()
 
-def length_update(name="queue"):
-    with sq3.connect("queue.db") as database:
-        c = database.cursor()
-        c.execute("select * from ?",(name,))
-        res = len(c.fetchall())
-        c.execute("update queue_manager set len = ? where name = '?'",(res,name))
-        database.commit()
-        c.close()
-    return res
 
 def length_get(name="queue"):
     with sq3.connect("queue.db") as database:
         c = database.cursor()
-        c.execute("select * from queue_manager where name = '?'",(name,))
+        c.execute("select * from queue_manager where name = ?",(str(name),))
         res = c.fetchone()[1]
         database.commit() #idk if i need to do this for read-only queries, but better safe than sorry?
         c.close()
+    if False:
+        print(f"length_get {name} {res}")
     return res
 
 def queue_write(content,name="queue"):
     with sq3.connect("queue.db") as database:
         c = database.cursor()
-        c.execute(f"{DB_INSERT(name)} {(length_get(name), str(content), 1, round(time.time()))}")
+        c.execute(f"{DB_INSERT(name)} (?, ?, 1, ?)",(int(length_get(name)),str(content),round(time.time())))
         #tuple to string formatting looks like this should work?
         database.commit()
         c.close()
     length_update(name) #should increase len
 
 def queue_pop(pos, name="queue"):
-    with sq3.connect("queue.db") as database:
-        c = database.cursor()
-        qLen = length_get(name)
-        if qLen > 0:
-            c.execute(f"select * from {name} where ID = {pos}")
-            res = str(c.fetchone()[1])
-            c.execute(f"delete from {name} where ID = {pos}")
-            for i in range(qLen-pos):
-                c.execute(f"update {name} set ID = {pos+i} where ID = {pos+i+1}")
-                #shift all entries above pos down by 1
-        else:
-            res = ""
-            print(f"empty return for pos {pos}")
-        database.commit()
-        c.close()
-    length_update(name)
-    return res
+    err = ""
+    res = ""
+    if san_bool(name):
+        with sq3.connect("queue.db") as database:
+            c = database.cursor()
+            qLen = length_get(name)
+            if qLen > 0:
+                c.execute(f"select * from {name} where ID = {pos}")
+                res = str(c.fetchone()[1])
+                c.execute(f"delete from {name} where ID = {pos}")
+                for i in range(qLen-pos):
+                    c.execute(f"update {name} set ID = {pos+i} where ID = {pos+i+1}")
+                    #shift all entries above pos down by 1
+            else:
+                res = ""
+                print(f"empty return for pos {pos}")
+            database.commit()
+            c.close()
+        length_update(name)
+    else:
+        err = "sanitise_name"
+    return (res,err)
 
 def queue_new(name,writePwd="", readPwd=""):
     global managerLen
@@ -146,28 +161,38 @@ def queue_new(name,writePwd="", readPwd=""):
     if err == "":
         queue_write(name,"info_queue")
     return((deletePwd,err))
-
+def manager_del_update(pos,c):
+    global managerLen
+    for i in range(managerLen-pos):
+                        c.execute(f"update queue_manager set ID = {pos+i} where ID = {pos+i+1}")
 def queue_delete(name, deletePwd):
+    global managerLen
     err = ""
     if not name in PROTECTED_NAMES:
-        with sq3.connect("queue.db") as database:
-            c = database.cursor()
-            c.execute("select * from queue_manager where name = '?'",(name,))
-            targetDb = c.fetchmany(1)[0]
-            if targetDb[3] == deletePwd:
-                c.execute("delete from queue_manager where name = '?' and delete_secret = '?'",(name,deletePwd))
-                c.execute("drop table ?",(name,))
-            else:
-                err = "incorrect_secret"
-            database.commit()
-            c.close()
-        return err
+        if san_bool(name):
+            with sq3.connect("queue.db") as database:
+                c = database.cursor()
+                c.execute("select * from queue_manager where name = ?",(str(name),))
+                targetDb = c.fetchmany(1)[0]
+                targetID = targetDb[0]
+                if targetDb[3] == deletePwd:
+                    c.execute("delete from queue_manager where name = ? and delete_secret = ?",(str(name),str(deletePwd)))
+                    c.execute(f"drop table {name}")
+                    manager_del_update(targetID,c)
+                    managerLen -= 1
+                else:
+                    err = "incorrect_secret"
+                database.commit()
+                c.close()
+            return err
+        else:
+            return "sanitise_name"
     else:
         return "protected_name"
 def q_info(name):
     with sq3.connect("queue.db") as database:
         c = database.cursor()
-        c.execute("select * from queue_manager where name = '?'",(name,))
+        c.execute("select * from queue_manager where name = ?",(str(name),))
         res = c.fetchmany(1)[0]
         database.commit()
         c.close()
